refactor: move platform type to common usage
