+++ src/tasks/commandTaskProvider.ts
@@ -1,9 +1,9 @@
 import * as vscode from 'vscode';
-import { TaskExecutionContext, Platform, ProjectType } from './tasksHelper';
+import { TaskExecutionContext, ProjectType } from './tasksHelper';
 import { TaskPseudoTerminal } from './taskPseudoTerminal';
 import { TaskHelper, Helpers } from './helpers';
 import { UserCancellation, handleInteractionError, InteractionError, checkLogin } from '../commands/common';
-import { LogLevel } from '../types/common';
+import { LogLevel, Platform } from '../types/common';
 import { CommandError } from '../common/utils';
 
 function getPlatform (task: TitaniumTaskBase): Platform {
