+++ src/tasks/helpers/index.ts
@@ -1,6 +1,5 @@
+import { Platform } from '../../types/common';
 import { TaskHelper } from './base';
-import { Platform } from '../tasksHelper';
-
 export * from './android';
 export * from './base';
 export * from './ios';
